Move/delete files for better structure
