+++ __tests__/campus-toggle.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CampusToggle from "../components/campus-toggle/campus-toggle.component";
+import CampusToggle from "../src/components/campus-toggle/campus-toggle.component";
 import renderer from "react-test-renderer";
 
 // Kind: Snapshot + Unit Tests
