+++ src/components/building-highlights/building-highlights.component.tsx
