+++ src/components/sliding-panel/sliding-panel.component.tsx
