+++ src/constants/building-information.data.ts
